refactor: consolidate authentication features into a single file
